move cloud storage writer to the corresponding package
