+++ downstreamadapter/sink/cloudstorage/ddl_worker.go
@@ -10,6 +10,7 @@
 // distributed under the License is distributed on an "AS IS" BASIS,
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 package cloudstorage
 
 import (
