+++ downstreamadapter/sink/cloudstorage/worker_test.go
@@ -10,7 +10,8 @@
 // distributed under the License is distributed on an "AS IS" BASIS,
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package writer
+
+package cloudstorage
 
 import (
 	"context"
